Move tests to single directory
